ci: improve validate package
