+++ test/validate/windows_validate.go
@@ -13,12 +13,22 @@ import (
 	"k8s.io/client-go/rest"
 )
 
+const (
+	cnsWinLabelSelector = "k8s-app=azure-cns-win"
+)
+
 var (
 	hnsEndPointCmd      = []string{"powershell", "-c", "Get-HnsEndpoint | ConvertTo-Json"}
 	hnsNetworkCmd       = []string{"powershell", "-c", "Get-HnsNetwork | ConvertTo-Json"}
 	azureVnetCmd        = []string{"powershell", "-c", "cat ../../k/azure-vnet.json"}
 	azureVnetIpamCmd    = []string{"powershell", "-c", "cat ../../k/azure-vnet-ipam.json"}
 	restartKubeProxyCmd = []string{"powershell", "Restart-service", "kubeproxy"}
+	cnsWinLocalCacheCmd = []string{
+		"powershell", "Invoke-WebRequest -Uri 127.0.0.1:10090/debug/ipaddresses",
+		"-Method Post -ContentType application/x-www-form-urlencoded",
+		"-Body \"{`\"IPConfigStateFilter`\":[`\"Assigned`\"]}\"",
+		"-UseBasicParsing | Select-Object -Expand Content",
+	}
 )
 
 var windowsChecksMap = map[string][]check{
@@ -28,7 +38,9 @@ var windowsChecksMap = map[string][]check{
 		{"azure-vnet-ipam", azureVnetIpamIps, privilegedLabelSelector, privilegedNamespace, azureVnetIpamCmd},
 	},
 	"cniv2": {
+		{"hns", hnsStateFileIps, privilegedLabelSelector, privilegedNamespace, hnsEndPointCmd},
 		{"azure-vnet", azureVnetIps, privilegedLabelSelector, privilegedNamespace, azureVnetCmd},
+		{"cns cache", cnsCacheStateFileIps, cnsWinLabelSelector, privilegedNamespace, cnsWinLocalCacheCmd},
 	},
 }
